# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Professional Services Marketplace API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
...
  /admin/projects:
    get:
      operationId: adminListProjects
      responses: { '200': { description: Paginated project moderation list }, '403': { $ref: '#/components/responses/Error' } }
...
  /me/projects:
    get:
      operationId: listMyProjects
      parameters:
        - { name: cursor, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        '200': { description: Cursor-paginated owned projects }
        '401': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
    post:
      operationId: createProject
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ProjectCreateRequest' } } }
      responses:
        '201': { description: Manual draft project created }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
...
components:
  schemas:
...
    ProjectBudgetType:
      type: string
      enum: [FIXED, RANGE, NEGOTIABLE, HOURLY]
    ProjectExperienceLevel:
      type: string
      enum: [BEGINNER, INTERMEDIATE, ADVANCED, EXPERT]
    ProjectBudget:
      type: object
      required: [type, currency]
      additionalProperties: false
      properties:
        type: { $ref: '#/components/schemas/ProjectBudgetType' }
        min_kopecks: { type: integer, format: int64, minimum: 1 }
        max_kopecks: { type: integer, format: int64, minimum: 1 }
        currency: { type: string, enum: [RUB] }
    ProjectCreateRequest:
      type: object
      required: [title, slug, description, budget, visibility]
      additionalProperties: false
      properties:
        category_id: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, minLength: 1, maxLength: 240, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
        description: { type: string, minLength: 1, maxLength: 15000 }
        budget: { $ref: '#/components/schemas/ProjectBudget' }
        deadline_at: { type: string, format: date-time }
        experience_level: { $ref: '#/components/schemas/ProjectExperienceLevel' }
        visibility: { type: string, enum: [PUBLIC, PRIVATE] }
        skill_ids: { type: array, maxItems: 30, uniqueItems: true, items: { type: string, format: uuid } }
        media_ids: { type: array, maxItems: 12, uniqueItems: true, items: { type: string, format: uuid } }
        source_draft_token: { type: string, minLength: 64, maxLength: 64 }
...
    Error:
      type: object
      properties:
        error: { type: object }
  responses:
    Error:
      description: API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }